+++ spring-cloud-tencent-plugin-starters/spring-cloud-starter-tencent-gateway-plugin/src/main/java/com/tencent/cloud/plugin/gateway/context/ContextRoutePredicateFactory.java
@@ -1,0 +1,78 @@
+/*
+ * Tencent is pleased to support the open source community by making spring-cloud-tencent available.
+ *
+ * Copyright (C) 2021 THL A29 Limited, a Tencent company. All rights reserved.
+ *
+ * Licensed under the BSD 3-Clause License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed
+ * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ * CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package com.tencent.cloud.plugin.gateway.context;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.function.Predicate;
+
+import org.springframework.cloud.gateway.handler.predicate.AbstractRoutePredicateFactory;
+import org.springframework.cloud.gateway.handler.predicate.GatewayPredicate;
+import org.springframework.web.server.ServerWebExchange;
+
+public class ContextRoutePredicateFactory extends AbstractRoutePredicateFactory<ContextRoutePredicateFactory.Config> {
+
+	public ContextRoutePredicateFactory() {
+		super(Config.class);
+	}
+
+	@Override
+	public Predicate<ServerWebExchange> apply(Config config) {
+		return new GatewayPredicate() {
+			@Override
+			public boolean test(ServerWebExchange exchange) {
+				// TODO: do path-rewriting , put to GATEWAY_PREDICATE_PATH_CONTAINER_ATTR
+				return true;
+			}
+
+			@Override
+			public Object getConfig() {
+				return config;
+			}
+
+			@Override
+			public String toString() {
+				return String.format("Config: %s", config);
+			}
+		};
+	}
+
+	@Override
+	public List<String> shortcutFieldOrder() {
+		return Arrays.asList("group");
+	}
+
+	public static class Config {
+		private String group;
+
+		public String getGroup() {
+			return group;
+		}
+
+		public void setGroup(String group) {
+			this.group = group;
+		}
+
+		@Override
+		public String toString() {
+			return "Config{" +
+					"group='" + group + '\'' +
+					'}';
+		}
+	}
+}