+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/instrument/feign/EnhancedFeignClient.java
@@ -97,6 +97,7 @@ public class EnhancedFeignClient implements Client {
 		try {
 			// Run pre enhanced plugins.
 			pluginRunner.run(EnhancedPluginType.Client.PRE, enhancedPluginContext);
+			startMillis = System.currentTimeMillis();
 
 			Response response = delegate.execute(request, options);
 			enhancedPluginContext.setDelay(System.currentTimeMillis() - startMillis);